Update sponsors and speaker pages
The sponsor and speaker pages have been updated a lot. The overviews
are now grids, font awesome is implemented for icons, the speaker page
for Ray Bogman has been added and the sponsors are implemnted via YAML.

Also the call for papers link in the head has been removed, because
this one has been closed for a while now.